Refactored package structure of growing map objects.
