+++ jsettlers.logic/src/jsettlers/logic/map/newGrid/objects/MapObjectsManager.java
@@ -41,11 +41,11 @@ import jsettlers.logic.objects.arrow.ArrowObject;
 import jsettlers.logic.objects.building.BuildingWorkAreaMarkObject;
 import jsettlers.logic.objects.building.ConstructionMarkObject;
 import jsettlers.logic.objects.building.InformableMapObject;
-import jsettlers.logic.objects.corn.Corn;
+import jsettlers.logic.objects.growing.Corn;
+import jsettlers.logic.objects.growing.tree.AdultTree;
+import jsettlers.logic.objects.growing.tree.Tree;
 import jsettlers.logic.objects.stack.StackMapObject;
 import jsettlers.logic.objects.stone.Stone;
-import jsettlers.logic.objects.tree.AdultTree;
-import jsettlers.logic.objects.tree.Tree;
 import jsettlers.logic.player.Player;
 import jsettlers.logic.timer.IScheduledTimerable;
 import jsettlers.logic.timer.RescheduleTimer;
