+++ jsettlers.logic/src/jsettlers/logic/objects/growing/tree/AdultTree.java
@@ -12,52 +12,32 @@
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
  * DEALINGS IN THE SOFTWARE.
  *******************************************************************************/
-package jsettlers.logic.objects.corn;
+package jsettlers.logic.objects.growing.tree;
 
-import jsettlers.common.mapobject.EMapObjectType;
 import jsettlers.common.position.ShortPoint2D;
-import jsettlers.logic.objects.GrowingObject;
 
 /**
- * This is a Corn on the map.
+ * This is a tree on the map, that's adult from the beginning.
  * 
  * @author Andreas Eberle
  * 
  */
-public final class Corn extends GrowingObject {
-	private static final long serialVersionUID = -7535441306083940418L;
-
-	public static final float GROWTH_DURATION = 7 * 60;
-	public static final float DECOMPOSE_DURATION = 3 * 60;
-	public static final float REMOVE_DURATION = 2 * 60;
+public final class AdultTree extends Tree {
+	private static final long serialVersionUID = 5956923025331740093L;
 
 	/**
-	 * Creates a new Corn.
+	 * Creates a new adult Tree.
 	 * 
 	 * @param grid
 	 */
-	public Corn(ShortPoint2D pos) {
-		super(pos, EMapObjectType.CORN_GROWING);
+	public AdultTree(ShortPoint2D pos) {
+		super(pos);
+		super.changeState();
 	}
 
 	@Override
 	protected float getGrowthDuration() {
-		return GROWTH_DURATION;
-	}
-
-	@Override
-	protected float getDecomposeDuration() {
-		return DECOMPOSE_DURATION;
-	}
-
-	@Override
-	protected EMapObjectType getDeadState() {
-		return EMapObjectType.CORN_DEAD;
-	}
-
-	@Override
-	protected EMapObjectType getAdultState() {
-		return EMapObjectType.CORN_ADULT;
+		return 0.01f;
 	}
 
 }
