+++ jsettlers.logic/src/jsettlers/logic/objects/growing/tree/Tree.java
@@ -12,32 +12,71 @@
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
  * DEALINGS IN THE SOFTWARE.
  *******************************************************************************/
-package jsettlers.logic.objects.tree;
+package jsettlers.logic.objects.growing.tree;
 
+import jsettlers.common.mapobject.EMapObjectType;
 import jsettlers.common.position.ShortPoint2D;
+import jsettlers.common.position.RelativePoint;
+import jsettlers.common.sound.ISoundable;
+import jsettlers.logic.objects.growing.GrowingObject;
 
 /**
- * This is a tree on the map, that's adult from the beginning.
+ * This is a tree on the map.
  * 
  * @author Andreas Eberle
  * 
  */
-public final class AdultTree extends Tree {
-	private static final long serialVersionUID = 5956923025331740093L;
+public class Tree extends GrowingObject implements ISoundable {
+	private static final long serialVersionUID = 8241068714975746824L;
+
+	public static final float GROWTH_DURATION = 7 * 60;
+	public static final float DECOMPOSE_DURATION = 2 * 60;
+
+	private static final RelativePoint[] BLOCKED = new RelativePoint[] { new RelativePoint(0, 0) };
+
+	private boolean soundPlayed;
 
 	/**
-	 * Creates a new adult Tree.
+	 * Creates a new Tree.
 	 * 
 	 * @param grid
 	 */
-	public AdultTree(ShortPoint2D pos) {
-		super(pos);
-		super.changeState();
+	public Tree(ShortPoint2D pos) {
+		super(pos, EMapObjectType.TREE_GROWING);
+	}
+
+	@Override
+	public RelativePoint[] getBlockedTiles() {
+		return BLOCKED;
 	}
 
 	@Override
 	protected float getGrowthDuration() {
-		return 0.01f;
+		return GROWTH_DURATION;
 	}
 
+	@Override
+	protected float getDecomposeDuration() {
+		return DECOMPOSE_DURATION;
+	}
+
+	@Override
+	public void setSoundPlayed() {
+		soundPlayed = true;
+	}
+
+	@Override
+	public boolean isSoundPlayed() {
+		return soundPlayed;
+	}
+
+	@Override
+	protected EMapObjectType getDeadState() {
+		return EMapObjectType.TREE_DEAD;
+	}
+
+	@Override
+	protected EMapObjectType getAdultState() {
+		return EMapObjectType.TREE_ADULT;
+	}
 }
